+++ crates/mzoon/src/helper.rs
@@ -1,9 +1,11 @@
 mod download;
 mod file_compressor;
 mod read_to_vec;
+mod try_into_string;
 mod visit_files;
 
 pub use download::download;
 pub use file_compressor::{BrotliFileCompressor, FileCompressor, GzipFileCompressor};
 pub use read_to_vec::{AsyncReadToVec, ReadToVec};
+pub use try_into_string::TryIntoString;
 pub use visit_files::visit_files;
